管理端version1.0over

+++ src/main/java/com/wounom/kaoyanircpadmin/service/TieWenService.java
@@ -15,7 +15,7 @@ public interface TieWenService {
 
     Result getCheckTiewen();
 
-    Result checkTiewen(int tiewenId, int status);
+    Result checkTiewen(Long tiewenId, int status);
 
     Result getTiewenByUser(int userId);
